backend/FinLens/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
pydantic
passlib
bcrypt
python-jose
python-multipart
mysql-connector-python>=8.0.0
python-dotenv
alembic
pandas
openpyxl
pytesseract
Pillow 
opencv-python 
scikit-learn 
joblib
numpy
requests

